# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FleetOps SADC API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /org:
    get:
      tags: [Org]
      operationId: getOrg
      summary: Get the current org
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Org'
    patch:
      tags: [Org]
      operationId: updateOrg
      summary: Update org settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Org'
  /org/members:
    get:
      tags: [Org]
      operationId: listOrgMembers
      summary: List org members
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrgMember'
    post:
      tags: [Org]
      operationId: inviteOrgMember
      summary: Invite a new org member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgMemberInvite'
      responses:
        '201':
          description: Invited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgMember'
  /org/members/{memberId}:
    patch:
      tags: [Org]
      operationId: updateOrgMember
      summary: Update a member's role
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgMemberUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgMember'
    delete:
      tags: [Org]
      operationId: removeOrgMember
      summary: Remove a member from the org
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Removed

  # ─── FX Rates ────────────────────────────────────────────────────────────
...
  /loads:
    get:
      tags: [Loads]
      operationId: listLoads
      summary: List loads for the org
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: truckId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Load'
    post:
      tags: [Loads]
      operationId: createLoad
      summary: Create a new load
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'
  /loads/{loadId}:
    get:
      tags: [Loads]
      operationId: getLoad
      summary: Get a single load
      parameters:
        - name: loadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'
    patch:
      tags: [Loads]
      operationId: updateLoad
      summary: Update load details or status
      parameters:
        - name: loadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'
    delete:
      tags: [Loads]
      operationId: deleteLoad
      summary: Delete a load
      parameters:
        - name: loadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
  /loads/{loadId}/assign:
    post:
      tags: [Loads]
      operationId: assignLoad
      summary: Assign a load to a truck
      parameters:
        - name: loadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadAssignment'
      responses:
        '200':
          description: Assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'
...
  /loads/{loadId}/invoice:
    post:
      tags: [Loads]
      operationId: generateInvoice
      summary: Generate invoice PDF for a delivered load
      parameters:
        - name: loadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'
  /loads/{loadId}/mark-paid:
    post:
      tags: [Loads]
      operationId: markInvoicePaid
      summary: Mark invoice as paid
      parameters:
        - name: loadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoicePaidUpdate'
      responses:
        '200':
          description: Marked paid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'

  # ─── Trips ────────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    Org:
      type: object
      required: [id, name, baseCurrency, invoiceTermsDays, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        country:
          type: [string, 'null']
        companyRegNo:
          type: [string, 'null']
        cbrtaOperatorNo:
          type: [string, 'null']
        vatNo:
          type: [string, 'null']
        address:
          type: [string, 'null']
        billingEmail:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        baseCurrency:
          type: string
        invoiceTermsDays:
          type: integer
        nextInvoiceSeq:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: [string, 'null']
          format: date-time
    OrgUpdate:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
        companyRegNo:
          type: string
        cbrtaOperatorNo:
          type: string
        vatNo:
          type: string
        address:
          type: string
        billingEmail:
          type: string
        phone:
          type: string
        baseCurrency:
          type: string
        invoiceTermsDays:
          type: integer
    OrgMember:
      type: object
      required: [id, orgId, userId, role, displayName, createdAt]
      properties:
        id:
          type: string
        orgId:
          type: string
        userId:
          type: string
        role:
          type: string
        displayName:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    OrgMemberInvite:
      type: object
      required: [role]
      properties:
        displayName:
          type: string
        phone:
          type: string
        role:
          type: string
    OrgMemberUpdate:
      type: object
      properties:
        role:
          type: string
        displayName:
          type: string
        phone:
          type: string

    # ─── FX Rate Schemas ─────────────────────────────────────────────────
...
    Load:
      type: object
      required: [id, orgId, consignmentNo, customer, originCity, destCity, status, currency, invoiceStatus, createdAt]
      properties:
        id:
          type: string
        orgId:
          type: string
        consignmentNo:
          type: string
        customer:
          type: string
        originCity:
          type: string
        destCity:
          type: string
        corridor:
          type: [string, 'null']
        borderPosts:
          type: array
          items:
            type: string
        distanceKm:
          type: [number, 'null']
        cargoDesc:
          type: [string, 'null']
        cargoWeightKg:
          type: [number, 'null']
        customsRef:
          type: [string, 'null']
        rate:
          type: [number, 'null']
        currency:
          type: string
        pickupDate:
          type: [string, 'null']
          format: date
        deliveredDate:
          type: [string, 'null']
          format: date
        truckId:
          type: [string, 'null']
        truckFleetNo:
          type: [string, 'null']
        status:
          type: string
        onTime:
          type: [boolean, 'null']
        podUrl:
          type: [string, 'null']
        invoiceNumber:
          type: [string, 'null']
        invoiceStatus:
          type: string
        invoiceDate:
          type: [string, 'null']
          format: date
        dueDate:
          type: [string, 'null']
          format: date
        paidDate:
          type: [string, 'null']
          format: date
        invoicePdfUrl:
          type: [string, 'null']
        daysOverdue:
          type: [integer, 'null']
        ratePerKm:
          type: [number, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: [string, 'null']
          format: date-time
    LoadInput:
      type: object
      required: [consignmentNo, customer, originCity, destCity, currency]
      properties:
        consignmentNo:
          type: string
        customer:
          type: string
        originCity:
          type: string
        destCity:
          type: string
        corridor:
          type: string
        borderPosts:
          type: array
          items:
            type: string
        distanceKm:
          type: number
        cargoDesc:
          type: string
        cargoWeightKg:
          type: number
        customsRef:
          type: string
        rate:
          type: number
        currency:
          type: string
        pickupDate:
          type: string
          format: date
    LoadUpdate:
      type: object
      properties:
        consignmentNo:
          type: string
        customer:
          type: string
        originCity:
          type: string
        destCity:
          type: string
        corridor:
          type: string
        borderPosts:
          type: array
          items:
            type: string
        distanceKm:
          type: number
        cargoDesc:
          type: string
        cargoWeightKg:
          type: number
        customsRef:
          type: string
        rate:
          type: number
        currency:
          type: string
        pickupDate:
          type: string
          format: date
        status:
          type: string
    LoadAssignment:
      type: object
      required: [truckId]
      properties:
        truckId:
          type: string
...
    InvoicePaidUpdate:
      type: object
      required: [paidDate]
      properties:
        paidDate:
          type: string
          format: date

    # ─── Trip Schemas ─────────────────────────────────────────────────────
...