# lib/api-zod/src/generated/types/orgMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FleetOps SADC API
 * OpenAPI spec version: 0.1.0
 */

export interface OrgMember {
  id: string;
  orgId: string;
  userId: string;
  role: string;
  /** @nullable */
  displayName: string | null;
  /** @nullable */
  phone?: string | null;
  createdAt: Date;
}